feat(auth): add register notification; move model
