+++ frontend/src/app/feature/auth/components/register/register.component.ts
@@ -1,8 +1,10 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 
 import { UsersService } from '../services/users.service';
 import { RegisterFormComponent } from './register-form/register-form.component';
-import { UserPostRequest } from '../services/users-api.model';
+import { NotificationService } from '@shared/services/notification.service';
+import { UserPostRequest } from '../models/users-api.model';
 
 @Component({
   selector: 'suo-register',
@@ -11,10 +13,23 @@ import { UserPostRequest } from '../services/users-api.model';
   styleUrl: './register.component.scss',
 })
 export class RegisterComponent {
-  constructor(private usersService: UsersService) {}
+  constructor(
+    private usersService: UsersService,
+    private notificationService: NotificationService,
+    private router: Router
+  ) {}
 
   async onFormSubmitted(data: UserPostRequest) {
-    const resp = await this.usersService.registerUser(data);
-    console.log(resp);
+    try {
+      const resp = await this.usersService.registerUser(data);
+      console.log(resp);
+      this.notificationService.openSnackBar(
+        'You have been registered. You can now log in.'
+      );
+      this.router.navigate(['auth', 'login']);
+    } catch (error) {
+      this.notificationService.openSnackBar('Failed to register');
+      throw error;
+    }
   }
 }
